Breaks jobs.py to submodule and fixes tests
